# service router predicates