--- a/pom.xml
+++ b/pom.xml
@@ -78,6 +78,7 @@
     <maven-release-plugin.version>2.5</maven-release-plugin.version>
     <maven-enforcer-plugin.version>3.0.0-M1</maven-enforcer-plugin.version>
     <jsr305.version>3.0.2</jsr305.version>
+    <jctools.version>2.1.2</jctools.version>
   </properties>
 
     <modules>
@@ -207,6 +208,11 @@
         <groupId>io.reactivex</groupId>
         <artifactId>rxjava</artifactId>
         <version>${rxjava.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.jctools</groupId>
+        <artifactId>jctools-core</artifactId>
+        <version>${jctools.version}</version>
       </dependency>
       <dependency>
         <groupId>com.google.code.findbugs</groupId>