--- a/services/pom.xml
+++ b/services/pom.xml
@@ -44,6 +44,11 @@
       <groupId>io.dropwizard.metrics</groupId>
       <artifactId>metrics-core</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.jctools</groupId>
+      <artifactId>jctools-core</artifactId>
+    </dependency>
+
 
     <!-- Jackson deps -->
     <dependency>