--- a/services/src/main/java/io/scalecube/services/registry/ServiceRegistryImpl.java
+++ b/services/src/main/java/io/scalecube/services/registry/ServiceRegistryImpl.java
@@ -4,12 +4,13 @@ import io.scalecube.services.ServiceEndpoint;
 import io.scalecube.services.ServiceReference;
 import io.scalecube.services.registry.api.ServiceRegistry;
 
+import org.jctools.maps.NonBlockingHashMap;
+
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
+import java.util.Map;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
@@ -18,8 +19,8 @@ import java.util.stream.Stream;
 public class ServiceRegistryImpl implements ServiceRegistry {
 
   // todo how to remove it (tags problem)?
-  private final ConcurrentMap<String, ServiceEndpoint> serviceEndpoints = new ConcurrentHashMap<>();
-  private final ConcurrentMap<String, List<ServiceReference>> referencesByQualifier = new ConcurrentHashMap<>();
+  private final Map<String, ServiceEndpoint> serviceEndpoints = new NonBlockingHashMap<>();
+  private final Map<String, List<ServiceReference>> referencesByQualifier = new NonBlockingHashMap<>();
 
   @Override
   public List<ServiceEndpoint> listServiceEndpoints() {