before

    Benchmark                                     (count)   Mode  Cnt        Score        Error  Units
    RouterBenchmarks.roundRobinRouterThroughput         1  thrpt    3  1065297.749 ± 740283.296  ops/s
    RouterBenchmarks.roundRobinRouterThroughput        10  thrpt    3   202341.851 ± 308850.328  ops/s
    RouterBenchmarks.roundRobinRouterThroughput       100  thrpt    3    22412.995 ±   1153.310  ops/s
    RouterBenchmarks.roundRobinRouterAverageTime        1   avgt    3     3927.878 ±    115.432  ns/op
    RouterBenchmarks.roundRobinRouterAverageTime       10   avgt    3    17477.033 ±    696.375  ns/op
    RouterBenchmarks.roundRobinRouterAverageTime      100   avgt    3   173403.971 ±   5737.451  ns/op

after

    Benchmark                                     (count)   Mode  Cnt        Score         Error  Units
    RouterBenchmarks.roundRobinRouterThroughput         1  thrpt    3  8893455.637 ± 1452938.551  ops/s
    RouterBenchmarks.roundRobinRouterThroughput        10  thrpt    3  9843796.291 ±  741562.485  ops/s
    RouterBenchmarks.roundRobinRouterThroughput       100  thrpt    3  8810162.492 ± 1903838.163  ops/s
    RouterBenchmarks.roundRobinRouterAverageTime        1   avgt    3      397.952 ±     132.824  ns/op
    RouterBenchmarks.roundRobinRouterAverageTime       10   avgt    3      452.489 ±      12.505  ns/op
    RouterBenchmarks.roundRobinRouterAverageTime      100   avgt    3      430.106 ±      46.226  ns/op

after applying NonBlockingHashMap

    Benchmark                                     (count)   Mode  Cnt         Score         Error  Units
    RouterBenchmarks.roundRobinRouterThroughput         1  thrpt    3  28091703.043 ± 5528210.067  ops/s
    RouterBenchmarks.roundRobinRouterThroughput        10  thrpt    3  29426425.413 ±  922160.465  ops/s
    RouterBenchmarks.roundRobinRouterThroughput       100  thrpt    3  29103953.007 ± 1231005.632  ops/s
    RouterBenchmarks.roundRobinRouterAverageTime        1   avgt    3       195.671 ±      27.186  ns/op
    RouterBenchmarks.roundRobinRouterAverageTime       10   avgt    3       142.309 ±       7.923  ns/op
    RouterBenchmarks.roundRobinRouterAverageTime      100   avgt    3       189.902 ±      18.439  ns/op